Refactor and decouple core sensor classes, reuse analyzed ktFiles from detekt
